Simplify code generation by using stateful output routines.
This replaces the "pure functional" output style, eliminating the need for
chaining semicolons and "next" continuation parameters.  Now when you call
routines like say, nl, put, and indent, the effects of those calls are going
directly into a file handle or other target as they run.  You are no longer
building up a lazy nested piece of data which is _later_ sent to an output
target.

The pure functional style still has its place in certain aspects of a program,
but for input and output it's easier just to cut to the chase and deal with the
machine on its own terms.  In those cases, "var" is your friend -- and it can
even create more opportunities for implementing routines directly in C, as I
did with the parsing primitives in stream.c in the Fexl project.

The good thing about "var" is that it is _controlled_ mutability, not the
haphazard and dangerous kind where you modify nested data structures in place.
In Fexl the data structures are still immutable, but you use a var to _point_
to a new version when you make a state change.  I think this strikes the right
balance between pure functions and routines with side effects.